[WIP] Synchronize tid allocation with datum insert
Logic here is to keep list of split tid allocation ranges. For example,
if a multi-insert allocates tids [1-1k, 1k-2k, 3k-4k] we note that there
is a hole between 2k-3k. Then when we later insert the datums into the
page, we split our stream and insert only the left-hand-side of the
split. Then we wait until it is the right time for the right-hand-side
to insert based on ZSRootDirItem.lasttidinserted.

Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>